b_up/requirements.txt
firebase-admin
python-dotenv
langchain
langchain-openai
Flask
Flask-Cors